sway: Migrate from i3blocks to waybar
The waybar configuration is a stripped-down version of the default. It
should display basically the same information as i3blocks did.
Additionally, the workspace names are now handled by waybar instead of
sway itself, as this allows the workspaces to stay in chronological
order e.g. when opening workspace 2 while workspace 5 is open,
workspace 2 should appear before workspace 5 in the status bar. When
workspace naming was handled by sway, this scenario would cause
workspace 2 to appear _after_ workspace 5 in the status bar.