waybar: Use custom updates script
There was a potential license incompatibility with the previous script,
and the old one felt hackish anyway. In order to avoid pango markup
embedded in the new script, the foreground color of the updates block
was changed back to a white, so say goodbye to the blue Arch logo there
(for now; it may come back).