alacritty: Fix "unknown terminal type" in remote sessions
Most devices don't have alacritty installed, causing them to not
recognize TERM=alacritty and default the terminal to builtin_ansi. This
is undesirable as it causes funky and colorless terminal output when
remotely connecting to other machines. Most machines recognize
xterm-256color as a valid terminal type, however, and setting TERM to
this loses no functionality on the client side.