vim: Use moonfly colorscheme
vim-plug was the chosen package manager primarily for its minimalism.
Although I'm not the greatest fan of Rust doc-comment contrast in this
colorscheme, it's the best I can find and it suits me very well.

Rust extended syntax was installed additionally because it aligns with
the purpose of making syntax highlighting better for myself.